scripts/requirements.txt
sentence-transformers==3.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
